requirements.txt
anthropic==0.7.0
numpy==1.24.3
pillow==10.2.0
matplotlib==3.8.2
opencv-python==4.8.0.74
nuscenes-devkit==1.1.10
pyquaternion==0.9.9
tqdm==4.66.1
typer==0.9.0
hydra-core==1.3.2
rich==13.7.0
pydantic==2.6.1